Move frontend tests to different file
They don't depend only on the LDAP.
