+++ tests/integration/test_hss_ldap.py
@@ -2,7 +2,6 @@ from functools import partial
 import unittest
 from unittest.mock import patch
 
-from flask import url_for
 import ldap3
 from ldap3.core.exceptions import LDAPPasswordIsMandatoryError, LDAPBindError
 
@@ -272,70 +271,5 @@ class GetTestCase(SimpleLdapUserTestBase):
         )
 
 
-class HssPgConfigMixin:
-    def create_app(self, *a, **kw):
-        conf = {
-            **kw.pop('additional_config', {}),
-            'HSS_CONNECTION_STRING': "postgresql://sipa:password@postgres:5432/",
-        }
-        return super().create_app(*a, additional_config=conf, **kw)
-
-
 class SimpleHssPgTestBase(HSSOneAccountFixture, HssPgTestBase):
     pass
-
-
-class HssAuthenticatedTestBase(HssPgConfigMixin, SimpleLdapTestBase):
-    def __init__(self, *a, **kw):
-        super().__init__(*a, **kw)
-        self.uid = list(self.fixtures.keys()).pop()
-        self.pw = self.fixtures[self.uid]['userPassword']
-
-    def login(self):
-        return self.client.post(
-            url_for('generic.login'),
-            data={'dormitory': 'hss',
-                  'username': self.uid,
-                  'password': self.pw}
-        )
-
-    def logout(self):
-        return self.client.get(
-            url_for('generic.logout')
-        )
-
-
-class HssAuthenticationTestCase(HssAuthenticatedTestBase):
-    def test_login_successful(self):
-        rv = self.login()
-        self.assertRedirects(rv, url_for('usersuite.usersuite'))
-
-    def test_logout_successful(self):
-        self.login()
-        rv = self.logout()
-        if rv.status_code == 401:
-            self.fail('Logout not permitted, probably because login failed')
-        self.assert_redirects(rv, url_for('generic.index'))
-
-
-class HssFrontendTestBase(HssAuthenticatedTestBase):
-    def setUp(self):
-        super().setUp()
-        self.login()
-
-    def tearDown(self):
-        self.logout()
-        super().tearDown()
-
-
-class HssPasswordChangeTestCase(HssFrontendTestBase):
-    def setUp(self):
-        super().setUp()
-        self.rv = self.client.post(
-            url_for('usersuite.usersuite_change_password'),
-            data={'username': self.uid, 'password': self.pw}
-        )
-        self.text = "Diese Funktion ist nicht verfügbar.".encode('utf-8')
-
-    def test_password_change_not_disallowed(self):
-        self.assertIn(self.text, self.rv.data)
